Move gazctl command to make gazctlcmd fully reusable
